17: Inline json mapping classes
Task-Url: https://github.com/exasol/bucketfs-java/issues/17
